backend/requirements.txt
fastapi
uvicorn
strawberry-graphql
transformers
torch